Adapt Milvus2.2 Channel checkpoint
