+++ states/show_segment.go
@@ -10,7 +10,6 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/milvus-io/birdwatcher/proto/v2.0/commonpb"
 	"github.com/milvus-io/birdwatcher/proto/v2.0/datapb"
-	"github.com/milvus-io/birdwatcher/proto/v2.0/internalpb"
 	"github.com/milvus-io/birdwatcher/proto/v2.0/querypb"
 	datapbv2 "github.com/milvus-io/birdwatcher/proto/v2.2/datapb"
 	"github.com/milvus-io/birdwatcher/storage"
@@ -116,73 +115,6 @@ func getLoadedSegmentsCmd(cli *clientv3.Client, basePath string) *cobra.Command 
 	cmd.Flags().String("format", "line", "segment display format")
 	cmd.Flags().Bool("detail", false, "flags indicating whether pring detail binlog info")
 	cmd.Flags().Int64("segment", 0, "segment id to filter with")
-	return cmd
-}
-
-func getCheckpointCmd(cli *clientv3.Client, basePath string) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "checkpoint",
-		Short:   "list checkpoint collection vchannels",
-		Aliases: []string{"checkpoints", "cp"},
-		Run: func(cmd *cobra.Command, args []string) {
-
-			collID, err := cmd.Flags().GetInt64("collection")
-			if err != nil {
-				fmt.Println(err.Error())
-				return
-			}
-
-			coll, err := getCollectionByID(cli, basePath, collID)
-			if err != nil {
-				fmt.Println("failed to get collection", err.Error())
-				return
-			}
-
-			for _, vchannel := range coll.GetVirtualChannelNames() {
-				segments, err := listSegments(cli, basePath, func(info *datapb.SegmentInfo) bool {
-					return info.CollectionID == collID && info.InsertChannel == vchannel
-				})
-				if err != nil {
-					fmt.Printf("fail to list segment for channel %s, err: %s\n", vchannel, err.Error())
-					continue
-				}
-				fmt.Printf("find segments to list checkpoint for %s, segment found %d\n", vchannel, len(segments))
-				var segmentID int64
-				var pos *internalpb.MsgPosition
-				for _, segment := range segments {
-					if segment.State != commonpb.SegmentState_Flushed &&
-						segment.State != commonpb.SegmentState_Growing &&
-						segment.State != commonpb.SegmentState_Flushing {
-						continue
-					}
-					// skip all empty segment
-					if segment.GetDmlPosition() == nil && segment.GetStartPosition() == nil {
-						continue
-					}
-					var segPos *internalpb.MsgPosition
-
-					if segment.GetDmlPosition() != nil {
-						segPos = segment.GetDmlPosition()
-					} else {
-						segPos = segment.GetStartPosition()
-					}
-
-					if pos == nil || segPos.GetTimestamp() < pos.GetTimestamp() {
-						pos = segPos
-						segmentID = segment.GetID()
-					}
-				}
-
-				if pos == nil {
-					fmt.Printf("vchannel %s position nil\n", vchannel)
-				} else {
-					t, _ := ParseTS(pos.GetTimestamp())
-					fmt.Printf("vchannel %s seek to %v, for segment ID:%d \n", vchannel, t, segmentID)
-				}
-			}
-		},
-	}
-	cmd.Flags().Int64("collection", 0, "collection id to filter with")
 	return cmd
 }
 
@@ -451,6 +383,7 @@ func printSegmentInfo(info *datapb.SegmentInfo, detailBinlog bool) {
 		countBinlogNum(info.Binlogs), countBinlogNum(info.Statslogs), countBinlogNum(info.Deltalogs))
 
 	if detailBinlog {
+		var binlogSize int64
 		fmt.Println("**************************************")
 		fmt.Println("Binlogs:")
 		sort.Slice(info.Binlogs, func(i, j int) bool {
@@ -465,6 +398,7 @@ func printSegmentInfo(info *datapb.SegmentInfo, detailBinlog bool) {
 				fmt.Printf("Log Size: %d \t Entry Num: %d\t TimeRange:%s-%s\n",
 					binlog.LogSize, binlog.EntriesNum,
 					tf.Format(tsPrintFormat), tt.Format(tsPrintFormat))
+				binlogSize += binlog.LogSize
 			}
 		}
